added empty app and pipeline
