+++ .gitlab-ci.yml
@@ -1,0 +1,17 @@
+include:
+  - file: "/build/node/template-node-npm.yml"
+    project: "platform/pipelines-templates"
+    ref: "master"
+  - file: "/build/node/template-node-npm-test-latest.yml"
+    project: "platform/pipelines-templates"
+    ref: "master"
+
+.docker-job:
+  variables:
+    COMMIT_SHA_FILEPATH: /etc/nginx/commit.sha
+
+default:
+  image: node:14.18.1-alpine
+
+variables:
+  IMAGE_NAME: mia_template_image_name_placeholder
